requirements.txt
amqp==5.0.5
billiard==3.6.3.0
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
idna==2.10
importlib-metadata==3.4.0
kombu==5.0.2
prompt-toolkit==3.0.16
python-dotenv==0.15.0
pytz==2021.1
requests==2.25.1
six==1.15.0
SQLAlchemy==1.3.23
typing-extensions==3.7.4.3
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
zipp==3.4.0
